+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] / 2024-12-13
+### Features
+- Create `Autodesk.Forge` internal project.
+### Autodesk.Forge
+- Copy `BucketsApi`, `ObjectsApi`, `TwoLeggedApi` from the original `forge-api-dotnet-client` repository.
+- Update deprecated `x-ads-region` to `region` when creating a bucket. [buckets-POST](https://aps.autodesk.com/en/docs/data/v2/reference/http/buckets-POST/) (Fix: #10)
+### Updated
+- Update `CreateBucketAsync` to work with `region`.
+### Tests
+- Add `BucketApi_CreateAndDelete_Region`
+
 ## [2.1.0] / 2024-12-12
 ### Updated
 - Remove obsolete endpoints and use `S3`.
@@ -32,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `Autodesk.Forge.Oss.Tests` project
 
 [vNext]: ../../compare/1.0.0...HEAD
+[3.0.0]: ../../compare/2.1.0...3.0.0
 [2.1.0]: ../../compare/2.0.0...2.1.0
 [2.0.0]: ../../compare/1.0.2...2.0.0
 [1.0.2]: ../../compare/1.0.1...1.0.2